requirements.txt
pyyaml
loguru
orjson
polars